Add rowref benchmarks
